G2P-776: Edit Options added using dynamic mixin
