+++ g2p_payment_cash/models/payment_manager.py
@@ -60,22 +60,3 @@ class G2PPaymentManagerCash(models.Model):
                     }
                 )
             batch.batch_has_completed = True
-
-
-class PaymentInherit(models.Model):
-    _inherit = "g2p.payment"
-
-    edit_css = fields.Html(
-        sanitize=False,
-        compute="_compute_css",
-    )
-
-    def _compute_css(self):
-        for rec in self:
-            # To Remove Edit Option
-            if rec.status == "paid":
-                rec.edit_css = (
-                    "<style>.o_form_button_edit {display: none !important;}</style>"
-                )
-            else:
-                rec.edit_css = False
