+++ g2p_program_reimbursement/models/programs.py
@@ -8,11 +8,6 @@ class G2PPrograms(models.Model):
 
     reimbursement_program_id = fields.Many2one("g2p.program")
 
-    edit_css = fields.Html(
-        sanitize=False,
-        compute="_compute_css",
-    )
-
     def open_eligible_beneficiaries_form(self):
         res = super(G2PPrograms, self).open_eligible_beneficiaries_form()
         if self.is_reimbursement_program:
