+++ g2p_programs/models/payment.py
@@ -11,6 +11,7 @@ _logger = logging.getLogger(__name__)
 class G2PPayment(models.Model):
     _name = "g2p.payment"
     _description = "Payment"
+    _inherit = ["disable.edit.mixin"]
     _order = "id desc"
 
     name = fields.Char(
@@ -93,6 +94,16 @@ class G2PPayment(models.Model):
 
     def send_payment(self):
         pass
+
+    def _compute_css(self):
+        for rec in self:
+            # To Remove Edit Option
+            if rec.status == "paid":
+                rec.edit_css = (
+                    "<style>.o_form_button_edit {display: none !important;}</style>"
+                )
+            else:
+                rec.edit_css = False
 
 
 class G2PPaymentBatch(models.Model):
