+++ g2p_programs/models/programs.py
@@ -10,7 +10,12 @@ _logger = logging.getLogger(__name__)
 
 
 class G2PProgram(models.Model):
-    _inherit = ["mail.thread", "mail.activity.mixin", "job.relate.mixin"]
+    _inherit = [
+        "mail.thread",
+        "mail.activity.mixin",
+        "job.relate.mixin",
+        "disable.edit.mixin",
+    ]
     _name = "g2p.program"
     _description = "Program"
     _order = "id desc"
@@ -521,3 +526,14 @@ class G2PProgram(models.Model):
         jobs = self.env["queue.job"].search([("model_name", "like", self._name)])
         related_jobs = jobs.filtered(lambda r: self in r.records.program_id)
         return [("id", "in", related_jobs.ids)]
+
+    def _compute_css(self):
+        # Overriding this function
+        for rec in self:
+            # To Remove Edit Option
+            if rec.state == "ended":
+                rec.edit_css = (
+                    "<style>.o_form_button_edit {display: none !important;}</style>"
+                )
+            else:
+                rec.edit_css = False
